re-estruturei os arquivos do home
