+++ marvel-hero-navegator-frontend/src/components/NavBar/index.jsx
